Move things in Refact.Apply to Refact.Internal
